requirements.txt
Django~=3.1.3
psycopg2
GitPython
